pyproject.toml
[project]
name = "scope-test-generator-2"
version = "0.1.0"
dependencies = [
    "scope-worker-base",
    "click==3.1",  # sample conflicting dependency
]

[project.entry-points.scope]
test-generator-2 = "scope_test_generator"
